requirements.txt
flask 
PyPDF2 
langchain 
langchain-community 
langchain-google-genai 
faiss-cpu 
google-generativeai
dotenv-python
tiktoken
sentence-transformers